+++ charts/core/src/Line/Line.tsx
@@ -2,6 +2,7 @@ import { useEffect } from 'react';
 import { colors } from '@lg-charts/colors';
 import { useSeriesContext } from '@lg-charts/series-provider';
 
+import { useIdAllocator } from '@leafygreen-ui/hooks';
 import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
 
 import { useChartContext } from '../ChartContext';
@@ -11,20 +12,26 @@ import { LineProps } from './Line.types';
 
 export function Line({ name, data }: LineProps) {
   const { theme } = useDarkMode();
-  const { chart } = useChartContext();
+  const {
+    chart: { addSeries, ready, removeSeries },
+  } = useChartContext();
   const { getSeriesIndex, isChecked } = useSeriesContext();
+
+  const id = useIdAllocator({ prefix: 'line' });
 
   const themedColors = colors[theme];
   const colorIndex = getSeriesIndex(name) % themedColors.length; // loop through colors if more lines than available colors
   const color = themedColors[colorIndex];
+
   const isVisible = isChecked(name);
 
   useEffect(() => {
-    if (!chart.ready) return;
+    if (!ready) return;
 
     if (isVisible) {
-      chart.addSeries({
+      addSeries({
         ...defaultLineOptions,
+        id,
         name,
         data,
         lineStyle: {
@@ -37,7 +44,7 @@ export function Line({ name, data }: LineProps) {
         },
       });
     } else {
-      chart.removeSeries(name);
+      removeSeries(id);
     }
 
     return () => {
@@ -45,11 +52,9 @@ export function Line({ name, data }: LineProps) {
        * Remove the series when the component unmounts to make sure the series
        * is removed when a `Line` is hidden.
        */
-      chart.removeSeries(name);
+      removeSeries(id);
     };
-    // FIXME:
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [chart.ready, name, data, isVisible, theme]);
+  }, [addSeries, color, data, id, isVisible, name, ready, removeSeries, theme]);
 
   return null;
 }