+++ chat/input-bar/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @lg-chat/input-bar
 
+## 7.0.15
+
+### Patch Changes
+
+- Updated dependencies [2d40eba2a]
+  - @leafygreen-ui/popover@13.0.11
+  - @leafygreen-ui/search-input@5.0.13
+
+## 7.0.14
+
+### Patch Changes
+
+- Updated dependencies [f9fa0fe83]
+  - @leafygreen-ui/badge@9.0.10
+  - @leafygreen-ui/emotion@4.1.1
+
 ## 7.0.13
 
 ### Patch Changes