+++ chat/title-bar/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @lg-chat/title-bar
 
+## 3.0.14
+
+### Patch Changes
+
+- Updated dependencies [f9fa0fe83]
+  - @leafygreen-ui/badge@9.0.10
+  - @leafygreen-ui/emotion@4.1.1
+
 ## 3.0.13
 
 ### Patch Changes